requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
langgraph-checkpoint-sqlite
google-genai
python-dotenv
pluggy-sdk
tavily-python
filetype
langchain-google-genai